Move event handling out of split tunnel module
